refactor(docs): link dx
