+++ apps/docs/content/components/link/block.ts
@@ -1,29 +1,4 @@
-const App = `import {Link} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <div className="flex gap-2">
-      <Link isBlock showAnchorIcon href="#" color="foreground">
-        Foreground
-      </Link>
-      <Link isBlock showAnchorIcon href="#" color="primary">
-        Primary
-      </Link>
-      <Link isBlock showAnchorIcon href="#" color="secondary">
-        Secondary
-      </Link>
-      <Link isBlock showAnchorIcon href="#" color="success">
-        Success
-      </Link>
-      <Link isBlock showAnchorIcon href="#" color="warning">
-        Warning
-      </Link>
-      <Link isBlock showAnchorIcon href="#" color="danger">
-        Danger
-      </Link>
-    </div>
-  );
-}`;
+import App from "./block.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
