+++ apps/docs/content/components/link/colors.ts
@@ -1,17 +1,4 @@
-const App = `import {Link} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <div className="flex gap-4">
-      <Link href="#" color="foreground">Foreground</Link>
-      <Link href="#" color="primary">Primary</Link>
-      <Link href="#" color="secondary">Secondary</Link>
-      <Link href="#" color="success">Success</Link>
-      <Link href="#" color="warning">Warning</Link>
-      <Link href="#" color="danger">Danger</Link>
-    </div> 
-  );
-}`;
+import App from "./colors.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
