+++ apps/docs/content/components/link/custom-impl.ts
@@ -1,68 +1,5 @@
-const App = `import {forwardRef} from "react";
-import {LinkIcon} from "@nextui-org/shared-icons";
-import {linkAnchorClasses} from "@nextui-org/theme";
-
-import {useLink} from "@nextui-org/react";
-
-const MyLink = forwardRef((props, ref) => {
-  const {
-    Component,
-    children,
-    showAnchorIcon,
-    anchorIcon = <LinkIcon className={linkAnchorClasses} />,
-    getLinkProps,
-  } = useLink({
-    ...props,
-    ref,
-  });
-
-  return (
-    <Component {...getLinkProps()}>
-      <>
-        {children}
-        {showAnchorIcon && anchorIcon}
-      </>
-    </Component>
-  );
-});
-
-MyLink.displayName = "MyLink";
-
-export default MyLink;`;
-
-const AppTs = `import {forwardRef} from "react";
-import {LinkIcon} from "@nextui-org/shared-icons";
-import {linkAnchorClasses} from "@nextui-org/theme";
-
-import {LinkProps, useLink} from "@nextui-org/react";
-
-export interface MyLinkProps extends LinkProps {}
-
-const MyLink = forwardRef<HTMLAnchorElement, MyLinkProps>((props, ref) => {
-  const {
-    Component,
-    children,
-    showAnchorIcon,
-    anchorIcon = <LinkIcon className={linkAnchorClasses} />,
-    getLinkProps,
-  } = useLink({
-    ...props,
-    ref,
-  });
-
-  return (
-    <Component {...getLinkProps()}>
-      <>
-        {children}
-        {showAnchorIcon && anchorIcon}
-      </>
-    </Component>
-  );
-});
-
-MyLink.displayName = "MyLink";
-
-export default MyLink;`;
+import App from "./custom-impl.raw.jsx?raw";
+import AppTs from "./custom-impl.raw.tsx?raw";
 
 const react = {
   "/App.jsx": App,
