+++ apps/docs/content/components/link/disabled.raw.jsx
@@ -1,0 +1,9 @@
+import {Link} from "@nextui-org/react";
+
+export default function App() {
+  return (
+    <Link isDisabled href="#">
+      Disabled Link
+    </Link>
+  );
+}
