+++ apps/docs/content/components/link/external.ts
@@ -1,21 +1,4 @@
-const App = `import {Link} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <div className="flex gap-4">
-      <Link isExternal href="https://github.com/nextui-org/nextui">
-        External Link
-      </Link>
-      <Link
-        isExternal
-        href="https://github.com/nextui-org/nextui"
-        showAnchorIcon
-      >
-        External Link Anchor
-      </Link>
-    </div>
-  );
-}`;
+import App from "./external.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
