+++ apps/docs/content/components/link/nextjs.ts
@@ -1,13 +1,4 @@
-const App = `import {Link} from "@nextui-org/react";
-import NextLink from "next/link";
-
-export default function App() {
-  return (
-    <Link href="/route" as={NextLink}>
-      Next.js Link
-    </Link>
-  );
-}`;
+import App from "./nextjs.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
