+++ apps/docs/content/components/link/underline.raw.jsx
@@ -1,0 +1,23 @@
+import {Link} from "@nextui-org/react";
+
+export default function App() {
+  return (
+    <div className="flex gap-4">
+      <Link href="#" underline="none">
+        None
+      </Link>
+      <Link href="#" underline="hover">
+        Hover
+      </Link>
+      <Link href="#" underline="always">
+        Always
+      </Link>
+      <Link href="#" underline="active">
+        Active
+      </Link>
+      <Link href="#" underline="focus">
+        Focus
+      </Link>
+    </div>
+  );
+}
