+++ apps/docs/content/components/link/usage.raw.jsx
@@ -1,0 +1,5 @@
+import {Link} from "@nextui-org/react";
+
+export default function App() {
+  return <Link href="#">Default Link</Link>;
+}
